Implemented chats feature (#49)
* Added chats list which shows all chats.
* Added a chat detail view.
* Added chat button to profiles.
* Integrated with the api.
* Added compact layout.

NOTE: there is a big issue which can only be fixed by adjusting the backend quite a bit. It is not possible in the current state to get a student's open chats. So they will have to manually activate each chat every time they log in. The messages will be saved but we will need to add another layer to the backend chats feature to make it possible to get all currently open chats.

---------

Co-authored-by: Leopold Lemmermann <leopold@lemmermann.com>